move link validation functions from service to controller
