+++ api/src/components/link/__tests__/createLink.test.ts
@@ -1,6 +1,6 @@
 import * as testUtils from 'lib/testUtils';
 
-import * as linkService from 'link/service';
+import * as linkController from 'link/controller';
 
 describe('Create links', () => {
     beforeEach(async () => {
@@ -127,52 +127,52 @@ describe('Create links', () => {
 describe('canLinkBeCreatedBetweenPublicationTypes', () => {
     describe('Should return true', () => {
         test('PROBLEM -> HYPOTHESIS', () => {
-            const result = linkService.canLinkBeCreatedBetweenPublicationTypes('PROBLEM', 'HYPOTHESIS');
+            const result = linkController.canLinkBeCreatedBetweenPublicationTypes('PROBLEM', 'HYPOTHESIS');
             expect(result).toEqual(true);
         });
 
         test('PROBLEM -> DATA', () => {
-            const result = linkService.canLinkBeCreatedBetweenPublicationTypes('PROBLEM', 'DATA');
+            const result = linkController.canLinkBeCreatedBetweenPublicationTypes('PROBLEM', 'DATA');
             expect(result).toEqual(true);
         });
 
         test('PROBLEM -> INTERPRETATION', () => {
-            const result = linkService.canLinkBeCreatedBetweenPublicationTypes('PROBLEM', 'INTERPRETATION');
+            const result = linkController.canLinkBeCreatedBetweenPublicationTypes('PROBLEM', 'INTERPRETATION');
             expect(result).toEqual(true);
         });
 
         test('PROBLEM -> PROBLEM', () => {
-            const result = linkService.canLinkBeCreatedBetweenPublicationTypes('PROBLEM', 'PROBLEM');
+            const result = linkController.canLinkBeCreatedBetweenPublicationTypes('PROBLEM', 'PROBLEM');
             expect(result).toEqual(true);
         });
 
         test('HYPOTHESIS -> PROBLEM', () => {
-            const result = linkService.canLinkBeCreatedBetweenPublicationTypes('HYPOTHESIS', 'PROBLEM');
+            const result = linkController.canLinkBeCreatedBetweenPublicationTypes('HYPOTHESIS', 'PROBLEM');
             expect(result).toEqual(true);
         });
 
         test('PROTOCOL -> HYPOTHESIS', () => {
-            const result = linkService.canLinkBeCreatedBetweenPublicationTypes('PROTOCOL', 'HYPOTHESIS');
+            const result = linkController.canLinkBeCreatedBetweenPublicationTypes('PROTOCOL', 'HYPOTHESIS');
             expect(result).toEqual(true);
         });
 
         test('DATA -> PROTOCOL', () => {
-            const result = linkService.canLinkBeCreatedBetweenPublicationTypes('DATA', 'PROTOCOL');
+            const result = linkController.canLinkBeCreatedBetweenPublicationTypes('DATA', 'PROTOCOL');
             expect(result).toEqual(true);
         });
 
         test('ANALYSIS -> DATA', () => {
-            const result = linkService.canLinkBeCreatedBetweenPublicationTypes('ANALYSIS', 'DATA');
+            const result = linkController.canLinkBeCreatedBetweenPublicationTypes('ANALYSIS', 'DATA');
             expect(result).toEqual(true);
         });
 
         test('INTERPRETATION -> ANALYSIS', () => {
-            const result = linkService.canLinkBeCreatedBetweenPublicationTypes('INTERPRETATION', 'ANALYSIS');
+            const result = linkController.canLinkBeCreatedBetweenPublicationTypes('INTERPRETATION', 'ANALYSIS');
             expect(result).toEqual(true);
         });
 
         test('REAL_WORLD_APPLICATION -> INTERPRETATION', () => {
-            const result = linkService.canLinkBeCreatedBetweenPublicationTypes(
+            const result = linkController.canLinkBeCreatedBetweenPublicationTypes(
                 'REAL_WORLD_APPLICATION',
                 'INTERPRETATION'
             );
@@ -180,28 +180,28 @@ describe('canLinkBeCreatedBetweenPublicationTypes', () => {
         });
 
         test('PEER_REVIEW -> INTERPRETATION', () => {
-            const result = linkService.canLinkBeCreatedBetweenPublicationTypes('PEER_REVIEW', 'INTERPRETATION');
+            const result = linkController.canLinkBeCreatedBetweenPublicationTypes('PEER_REVIEW', 'INTERPRETATION');
             expect(result).toEqual(true);
         });
 
         test('PEER_REVIEW -> PROBLEM', () => {
-            const result = linkService.canLinkBeCreatedBetweenPublicationTypes('PEER_REVIEW', 'PROBLEM');
+            const result = linkController.canLinkBeCreatedBetweenPublicationTypes('PEER_REVIEW', 'PROBLEM');
             expect(result).toEqual(true);
         });
     });
     describe('Should return false', () => {
         test('PEER_REVIEW -> PEER_REVIEW', () => {
-            const result = linkService.canLinkBeCreatedBetweenPublicationTypes('PEER_REVIEW', 'PEER_REVIEW');
+            const result = linkController.canLinkBeCreatedBetweenPublicationTypes('PEER_REVIEW', 'PEER_REVIEW');
             expect(result).toEqual(false);
         });
 
         test('INTERPRETATION -> PROBLEM', () => {
-            const result = linkService.canLinkBeCreatedBetweenPublicationTypes('INTERPRETATION', 'PROBLEM');
+            const result = linkController.canLinkBeCreatedBetweenPublicationTypes('INTERPRETATION', 'PROBLEM');
             expect(result).toEqual(false);
         });
 
         test('PROTOCOL -> DATA', () => {
-            const result = linkService.canLinkBeCreatedBetweenPublicationTypes('PROTOCOL', 'DATA');
+            const result = linkController.canLinkBeCreatedBetweenPublicationTypes('PROTOCOL', 'DATA');
             expect(result).toEqual(false);
         });
     });
